+++ src/components/button/button_group/button_group.tsx
@@ -91,6 +91,10 @@ export type OuiButtonGroupProps = CommonProps & {
    */
   type?: 'single' | 'multi';
   /**
+   * Determines if the buttons should be horizontal oriented or vertical oriented
+   */
+  orientation?: 'horizontal' | 'vertical';
+  /**
    * An array of #OuiButtonGroupOptionProps
    */
   options: OuiButtonGroupOptionProps[];
@@ -155,6 +159,7 @@ export const OuiButtonGroup: FunctionComponent<Props> = ({
   onChange,
   options = [],
   type = 'single',
+  orientation = 'horizontal',
   ...rest
 }) => {
   // Compressed style can't support `ghost` color because it's more like a form field than a button
@@ -173,6 +178,7 @@ export const OuiButtonGroup: FunctionComponent<Props> = ({
     {
       'ouiButtonGroup--fullWidth': isFullWidth,
       'ouiButtonGroup--isDisabled': isDisabled,
+      'ouiButtonGroup--vertical': orientation === 'vertical',
     },
     className
   );